Extract topics related logic to the hook
